refactor: move channels_v2/api_channel.py to channels/api_channel.py
